language and attributes on fenced code blocks
Fenced code blocks can have a language given after the ```.
In GFM this is used to select code highlighting but we use it here
to do code execution magic.

Also included the possibility of an option string, which is as yet
unused. Format is

```language options
print "this is the code"
```