update regex for code blocks
Move away from consuming any number of blank lines before the code block
to forcing a single blank line or the start of the file.

i.e. code blocks have to be like

```code
x = 1
```

blah blah blah.

and not like
```code
x = 1
```
blah blah blah.

This is consistent with other markdown specs e.g. pandoc.

This required implementing a pre processor for text blocks (like we
already have for code blocks), which is currently only used to strip
excess blank lines from the start and end of the markdown text (which
are left there because of the new code regex).