remove dependence on notebook
The notebook is only required to use the ContentsManager, so remove it
to reduce the install requirements.

Anyone that is going to use this is very likely to be doing so through
the web interface in which case they have the notebook anyway.

However, we can't guarantee that they have upgraded from IPython3 so we
need to catch this case when importing the FileContentsManager.